requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
requests==2.31.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
